fixed npc dialogue triggering and positioning!

+++ systems/MouseInputSystem.js
@@ -222,8 +222,8 @@ export class MouseInputSystem {
             const dy = npcPos.y - playerPos.y;
             const distance = Math.sqrt(dx * dx + dy * dy) / this.TILE_SIZE;
             const interactRange = 2;
-
-            if (isClick || distance <= interactRange) {
+            // removed isClick || from this conditional to stop opening when range is too far.
+            if ( isClick && distance <= interactRange) {
                 if (!this.hasInteractedWithNPC) {
                     this.entityManager.removeComponentFromEntity('player', 'MouseTarget');
                     this.eventBus.emit('StopMovement', { entityId: 'player' });
